Add files via upload
I wasn't abel to download the right data that @Emanuel put on slack... and my enrichment analysis is not working. Unfortunately I did not find the time to solve all these problems and therefore submitt an incomplete html... but I tought it is still better to upload something than nothing at all!